requirements.txt
mysql.connector
database
datetime